Cast type for precision

+++ src/utils/utils.py
@@ -395,7 +395,8 @@ class HmsPredictionWriter(BasePredictionWriter):
     def write_on_batch_end(
         self, trainer, pl_module, prediction, batch_indices, batch, batch_idx, dataloader_idx
     ):
-        prediction = F.softmax(prediction, dim=1)
+        # Increase precision
+        prediction = F.softmax(prediction.to(torch.float32), dim=1)
         self.preds['eeg_id'].append(batch['meta']['eeg_id'])
         self.preds['prediction'].append(prediction.detach().cpu().numpy())
